+++ projects/example-app/src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AuthGuard } from '@example-app/auth/services/auth-guard.service';
+import { AuthGuard } from '@example-app/auth/services';
 import { NotFoundPageComponent } from '@example-app/core/containers/not-found-page.component';
 
 export const routes: Routes = [
