+++ projects/example-app/src/app/auth/containers/index.ts
@@ -1,0 +1,1 @@
+export * from './login-page.component';
