+++ projects/example-app/src/app/auth/effects/index.ts
@@ -1,0 +1,1 @@
+export * from './auth.effects';
